update modes to mix nested
