+++ tests/test_function_calls.py
@@ -1,4 +1,3 @@
-
 import pytest
 from pydantic import BaseModel
